pyproject.toml
[project]
name = "solvit"
version = "0.1.0"
description = "Здесь собраны зависимости по проекту, который содержит все Pet-проекты с платформы Solvit"
authors = [
    {name = "Netrebin D D"}
]
license = {text = "-"}
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "websockets (>=16.0,<17.0)",
    "aiogram (>=3.26.0,<4.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "fastapi (>=0.135.2,<0.136.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-order (>=1.3.0,<2.0.0)",
    "asgi-lifespan (==2.*)",
    "tox (>=4.53.0,<5.0.0)",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "python-dotenv>=1.2.2",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
package-mode = false

[tool.poetry.scripts]
test_api = "pytest tests/api_tracker/main.py -v"
test_quiz = "pytest tests/quiz_platform/main.py -v"
test_all = "pytest tests/ -v"
[dependency-groups]
dev = [
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pytest-cov>=7.1.0",
    "pytest-rerunfailures (>=16.1,<17.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)",
]

[tool.uv]
package = false
